requirements.txt
django-cors-headers==3.13.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
httpie==3.2.1
itypes==1.2.0
jinja2==3.1.2
pillow==9.3.0
pip-chill==1.0.1
pysocks==1.7.1
tzdata==2022.6
uritemplate==4.1.1
